[FE] create react loader (suspense) view
