+++ client/src/components/DarkModeToggleButton.jsx
@@ -4,26 +4,11 @@
  * Part of Silesian University of Technology project.
  * Created only for learning purposes.
  */
-import { useEffect, useState } from 'react';
 import { MdNightlightRound, MdOutlineWbSunny } from 'react-icons/md';
+import useDarkMode from '../hooks/useDarkMode';
 
 const DarkModeToggleButton = () => {
-  const [darkMode, setDarkMode] = useState(() => {
-    return JSON.parse(window.localStorage.getItem('darkMode')) || false;
-  });
-
-  useEffect(() => {
-    const setDarkModeClass = isDarkMode => {
-      const htmlElement = document.querySelector('html');
-      if (isDarkMode) {
-        htmlElement.classList.add('dark');
-      } else {
-        htmlElement.classList.remove('dark');
-      }
-    };
-    setDarkModeClass(darkMode);
-    window.localStorage.setItem('darkMode', JSON.stringify(darkMode));
-  }, [darkMode]);
+  const [darkMode, setDarkMode] = useDarkMode();
 
   const toggleDarkMode = () => {
     setDarkMode(!darkMode);
